+++ spec/spec_helper.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-require "flourish"
-
-RSpec.configure do |config|
-  config.order = :random
-  Kernel.srand config.seed
-end